Search by keyword in the employee list and configure some API from the server and enhance UI
