+++ .env
@@ -1,2 +1,3 @@
 VITE_API_BASE_URL=http://49.157.74.3/access_token/api
-VITE_API_BASE_URL_SPRING=http://192.168.120.89:8086/api/v1/
+VITE_API_BASE_URL_SPRING=http://192.168.110.28:925/api/v1
+VITE_API_BASE_URL_PIS_PIC=http://49.157.74.3/pis/public
