+++ src/api/laboratory.ts
@@ -4,6 +4,14 @@ export function retrieveAllLabResults(params: {} = {}) {
   return req.get("/labresults", params);
 }
 
+export function employeeLaborartoryResult(params: {} = {}) {
+  return req.get(`/employeeLaboratoryResult/${params.userid}`, params);
+}
+
+export function createLaboratoryResult(params: {} = {}) {
+  return req.post(`/labresults`,params);
+}
+
