+++ src/api/laboratory_req.ts
@@ -48,9 +48,9 @@ Axios.interceptors.response.use(
       }
     } else if (code) {
       if (code === 401) {
-        router.push({
-          path: "/login",
-        });
+        // router.push({
+        //   path: "/login",
+        // });
       } else if (code === 403) {
       } else {
         console.log(error.response.data.message);
