+++ src/layout/admin/Footer.vue
@@ -1,0 +1,75 @@
+<template>
+    <footer class="pt-4">
+        <div class="w-full px-6 mx-auto">
+        <div class="flex flex-wrap items-center -mx-3 lg:justify-between">
+            <div class="w-full max-w-full px-3 mt-0 mb-6 shrink-0 lg:mb-0 lg:w-1/2 lg:flex-none">
+            <div class="leading-normal text-center text-sm text-slate-500 lg:text-left">
+                ©
+                2022
+                made with <i class="fa fa-heart"></i> by
+                <a href="https://www.creative-tim.com" class="font-semibold text-slate-700" target="_blank">Creative Tim</a>
+                for a better web.
+            </div>
+            </div>
+            <div class="w-full max-w-full px-3 mt-0 shrink-0 lg:w-1/2 lg:flex-none">
+            <ul class="flex flex-wrap justify-center pl-0 mb-0 list-none lg:justify-end">
+                <li class="nav-item">
+                <a href="https://www.creative-tim.com" class="block px-4 pt-0 pb-1 font-normal transition-colors ease-soft-in-out text-sm text-slate-500" target="_blank">Creative Tim</a>
+                </li>
+                <li class="nav-item">
+                <a href="https://www.creative-tim.com/presentation" class="block px-4 pt-0 pb-1 font-normal transition-colors ease-soft-in-out text-sm text-slate-500" target="_blank">About Us</a>
+                </li>
+                <li class="nav-item">
+                <a href="https://creative-tim.com/blog" class="block px-4 pt-0 pb-1 font-normal transition-colors ease-soft-in-out text-sm text-slate-500" target="_blank">Blog</a>
+                </li>
+                <li class="nav-item">
+                <a href="https://www.creative-tim.com/license" class="block px-4 pt-0 pb-1 pr-0 font-normal transition-colors ease-soft-in-out text-sm text-slate-500" target="_blank">License</a>
+                </li>
+            </ul>
+            </div>
+        </div>
+        </div>
+    </footer>
+</template>
+
+<style scoped>
+    /* Fonts and icons */
+    @import 'https://fonts.googleapis.com/css?family=Open+Sans:300,400,600,700';
+    /* Font Awesome Icons */
+    @import 'https://kit.fontawesome.com/42d5adcbca.js';
+    /* Nucleo Icons */
+    @import '@/style/assets/css/nucleo-icons.css';
+    @import '@/style/assets/css/nucleo-svg.css';
+    /* Main Styling */
+    @import '@/style/assets/css/soft-ui-dashboard-tailwind.css';
+</style>
+
+<style lang="scss">
+  .pagination-container {
+    display: flex;
+    column-gap: 10px;
+  }
+  .paginate-buttons {
+    height: 40px;
+    width: 40px;
+    border-radius: 20px;
+    cursor: pointer;
+    background-color: rgb(242, 242, 242);
+    border: 1px solid rgb(217, 217, 217);
+    color: black;
+  }
+  .paginate-buttons:hover {
+    background-color: #d8d8d8;
+  }
+  .active-page {
+    background-color: #3498db;
+    border: 1px solid #3498db;
+    color: white;
+  }
+  .active-page:hover {
+    background-color: #2988c8;
+  }
+  .pagination-padding {
+    padding: 20px;
+  }
+</style>
