+++ src/layout/admin/SideBar.vue
@@ -1,8 +1,9 @@
 <script setup lang="ts">
     import { ref, onMounted } from "vue"
     import { S } from "@/utils"
-    import { useRoute, useRouter } from "vue-router"
+    import { useRouter } from "vue-router"
     import { useTokenStore } from "@/stores"
+    import { getUserProfile } from "@/api/auth"
 
     import LogoCt from "@/style/assets/img/team-2.jpg"
 
@@ -10,14 +11,25 @@
     const tokenStore = useTokenStore()
     const router = useRouter()
     const router_path = router.currentRoute.value.path
+    const user_profile = ref({ fname : "", lname : "" })
+    
+    onMounted(() => {
+      _getUserProfile()
+    })
 
+    const _getUserProfile = async () => {
+        user_profile.value = await getUserProfile()
+    }
+
+    console.log(user_profile.value)
+    
     const handleClickSignOut = () => {
         // S.delete('authToken')
         S.deleteAll(true);
         tokenStore.dispatch("");
 
         router.push({
-            path: "/login",
+            path: "/",
         });
     };
 
@@ -29,7 +41,7 @@
         <i class="absolute top-0 right-0 hidden p-4 opacity-50 cursor-pointer fas fa-times text-slate-400 xl:hidden" sidenav-close></i>
         <router-link to="/admin" class="block px-8 py-6 m-0 text-sm whitespace-nowrap text-slate-700" target="_blank">
           <img :src="logo_ct" class="inline h-full max-w-full transition-all duration-200 ease-nav-brand max-h-8" alt="main_logo" />
-          <span class="ml-1 font-semibold transition-all duration-200 ease-nav-brand">Soft UI Dashboard</span>
+          <span class="ml-1 font-semibold transition-all duration-200 ease-nav-brand">{{ user_profile.fname+" "+user_profile.lname }}</span>
         </router-link>
       </div>
 
@@ -59,28 +71,28 @@
           </li>
 
           <li class="mt-0.5 w-full">
-            <router-link :class="{ 'shadow-soft-xl rounded-lg bg-white font-semibold text-slate-700': router_path === '/table' }" to="/table" class="py-2.7 text-sm ease-nav-brand my-0 mx-4 flex items-center whitespace-nowrap px-4 transition-colors">
-              <div :class="{ 'bg-gradient-to-tl from-purple-700 to-pink-500 ': router_path === '/table' }" class="shadow-soft-2xl mr-2 flex h-8 w-8 items-center justify-center rounded-lg bg-white bg-center stroke-0 text-center xl:p-2.5">
+            <router-link :class="{ 'shadow-soft-xl rounded-lg bg-white font-semibold text-slate-700': router_path === '/laboratory' }" to="/laboratory" class="py-2.7 text-sm ease-nav-brand my-0 mx-4 flex items-center whitespace-nowrap px-4 transition-colors">
+              <div :class="{ 'bg-gradient-to-tl from-purple-700 to-pink-500 ': router_path === '/laboratory' }" class="shadow-soft-2xl mr-2 flex h-8 w-8 items-center justify-center rounded-lg bg-white bg-center stroke-0 text-center xl:p-2.5">
                 <svg width="12px" height="12px" viewBox="0 0 42 42" version="1.1" xmlns="http://www.w3.org/2000/svg" xmlns:xlink="http://www.w3.org/1999/xlink">
                   <title>office</title>
                   <g stroke="none" stroke-width="1" fill="none" fill-rule="evenodd">
                     <g transform="translate(-1869.000000, -293.000000)" fill="#FFFFFF" fill-rule="nonzero">
                       <g transform="translate(1716.000000, 291.000000)">
                         <g transform="translate(153.000000, 2.000000)">
-                          <path :class="{ 'fill-slate-800': router_path !== '/table' }" class="opacity-60" d="M12.25,17.5 L8.75,17.5 L8.75,1.75 C8.75,0.78225 9.53225,0 10.5,0 L31.5,0 C32.46775,0 33.25,0.78225 33.25,1.75 L33.25,12.25 L29.75,12.25 L29.75,3.5 L12.25,3.5 L12.25,17.5 Z"></path>
-                          <path :class="{ 'fill-slate-800': router_path !== '/table' }" d="M40.25,14 L24.5,14 C23.53225,14 22.75,14.78225 22.75,15.75 L22.75,38.5 L19.25,38.5 L19.25,22.75 C19.25,21.78225 18.46775,21 17.5,21 L1.75,21 C0.78225,21 0,21.78225 0,22.75 L0,40.25 C0,41.21775 0.78225,42 1.75,42 L40.25,42 C41.21775,42 42,41.21775 42,40.25 L42,15.75 C42,14.78225 41.21775,14 40.25,14 Z M12.25,36.75 L7,36.75 L7,33.25 L12.25,33.25 L12.25,36.75 Z M12.25,29.75 L7,29.75 L7,26.25 L12.25,26.25 L12.25,29.75 Z M35,36.75 L29.75,36.75 L29.75,33.25 L35,33.25 L35,36.75 Z M35,29.75 L29.75,29.75 L29.75,26.25 L35,26.25 L35,29.75 Z M35,22.75 L29.75,22.75 L29.75,19.25 L35,19.25 L35,22.75 Z"></path>
+                          <path :class="{ 'fill-slate-800': router_path !== '/laboratory' }" class="opacity-60" d="M12.25,17.5 L8.75,17.5 L8.75,1.75 C8.75,0.78225 9.53225,0 10.5,0 L31.5,0 C32.46775,0 33.25,0.78225 33.25,1.75 L33.25,12.25 L29.75,12.25 L29.75,3.5 L12.25,3.5 L12.25,17.5 Z"></path>
+                          <path :class="{ 'fill-slate-800': router_path !== '/laboratory' }" d="M40.25,14 L24.5,14 C23.53225,14 22.75,14.78225 22.75,15.75 L22.75,38.5 L19.25,38.5 L19.25,22.75 C19.25,21.78225 18.46775,21 17.5,21 L1.75,21 C0.78225,21 0,21.78225 0,22.75 L0,40.25 C0,41.21775 0.78225,42 1.75,42 L40.25,42 C41.21775,42 42,41.21775 42,40.25 L42,15.75 C42,14.78225 41.21775,14 40.25,14 Z M12.25,36.75 L7,36.75 L7,33.25 L12.25,33.25 L12.25,36.75 Z M12.25,29.75 L7,29.75 L7,26.25 L12.25,26.25 L12.25,29.75 Z M35,36.75 L29.75,36.75 L29.75,33.25 L35,33.25 L35,36.75 Z M35,29.75 L29.75,29.75 L29.75,26.25 L35,26.25 L35,29.75 Z M35,22.75 L29.75,22.75 L29.75,19.25 L35,19.25 L35,22.75 Z"></path>
                         </g>
                       </g>
                     </g>
                   </g>
                 </svg>
               </div>
-              <span class="ml-1 duration-300 opacity-100 pointer-events-none ease-soft">Tables</span>
+              <span class="ml-1 duration-300 opacity-100 pointer-events-none ease-soft">Laboratory</span>
             </router-link>
           </li>
 
-          <!-- <li class="mt-0.5 w-full">
-            <a class="py-2.7 text-sm ease-nav-brand my-0 mx-4 flex items-center whitespace-nowrap px-4 transition-colors" href="./pages/billing.html">
+          <li class="mt-0.5 w-full">
+            <a class="py-2.7 text-sm ease-nav-brand my-0 mx-4 flex items-center whitespace-nowrap px-4 transition-colors" href="#" @click="handleClickSignOut">
               <div class="shadow-soft-2xl mr-2 flex h-8 w-8 items-center justify-center rounded-lg bg-white bg-center fill-current stroke-0 text-center xl:p-2.5">
                 <svg width="12px" height="12px" viewBox="0 0 43 36" version="1.1" xmlns="http://www.w3.org/2000/svg" xmlns:xlink="http://www.w3.org/1999/xlink">
                   <title>credit-card</title>
@@ -96,10 +108,11 @@
                   </g>
                 </svg>
               </div>
-              <span class="ml-1 duration-300 opacity-100 pointer-events-none ease-soft">Billing</span>
+              <span class="ml-1 duration-300 opacity-100 pointer-events-none ease-soft">Sign Out</span>
             </a>
           </li>
 
+          <!-- 
           <li class="mt-0.5 w-full">
             <a class="py-2.7 text-sm ease-nav-brand my-0 mx-4 flex items-center whitespace-nowrap px-4 transition-colors" href="./pages/virtual-reality.html">
               <div class="shadow-soft-2xl mr-2 flex h-8 w-8 items-center justify-center rounded-lg bg-white bg-center stroke-0 text-center xl:p-2.5">
