+++ src/views/Home.vue
@@ -1,37 +1,6 @@
 <script setup lang="ts">
     import '@/style/index.css'
     import Nav from '../layout/Nav.vue'
-    import { ref, onMounted } from "vue"
-    import { useRoute, useRouter } from "vue-router";
-    import { S } from "@/utils";
-    import { useTokenStore } from "@/stores";
-    import { getUserProfile } from "@/api/auth";
-
-    const router = useRouter();
-    const route = useRoute();
-    const tokenStore = useTokenStore();
-
-    const showState = ref(false)
-
-    onMounted(() => {
-        _getUserProfile()
-    })
-
-    const _getUserProfile = async () => {
-        const response = await getUserProfile() 
-        console.log(response)
-    }
-
-    // 退出
-    const handleClickSignOut = () => {
-        // S.delete('authToken')
-        S.deleteAll(true);
-        tokenStore.dispatch("");
-
-        router.push({
-            path: "/login",
-        });
-    };
 </script>
 <template>
     <!-- <Nav></Nav> -->
